# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Subcults API
  description: |
    Privacy-first platform for mapping underground music communities.
    Combines real-time location awareness, live audio streaming, and trust-based discovery.

    ## Authentication
    JWT-based authentication with two token types:
    - **Access tokens**: 15-minute expiry, includes `did` (Decentralized Identifier) claim
    - **Refresh tokens**: 7-day expiry, cookie-only

    Pass the access token via `Authorization: Bearer <token>` header or the `access_token` cookie.

    ## Privacy
    All public location data uses geohash-based jitter to prevent precise location tracking.
    Precise coordinates are only available when `allow_precise` consent is granted.

    ## Rate Limiting
    Per-endpoint rate limits are enforced. When exceeded, the API returns `429 Too Many Requests`
    with a `Retry-After` header.
  version: 0.1.0
  contact:
    name: Subcults
  license:
    name: MIT
servers:
  - url: /
    description: Current deployment
...
paths:
...
  /internal/stripe:
    post:
      operationId: handleStripeWebhook
      tags: [Payments]
      summary: Stripe webhook handler
      description: Receives Stripe events (checkout.session.completed, payment_intent.succeeded, etc.). Verified via Stripe-Signature header.
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Invalid signature or payload

  # ── Utility ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT access token. Include as Authorization: Bearer <token> header or access_token cookie.'
...